requirements.txt
adal==1.2.7
altgraph==0.17
appdirs==1.4.4
atomicwrites==1.4.0
attrs==21.2.0
azure==4.0.0
azure-applicationinsights==0.1.0
azure-batch==4.1.3
azure-common==1.1.27
azure-cosmosdb-nspkg==2.0.2
azure-cosmosdb-table==1.0.6
azure-datalake-store==0.0.52
azure-eventgrid==1.3.0
azure-graphrbac==0.40.0
azure-keyvault==1.1.0
azure-loganalytics==0.1.0
azure-mgmt==4.0.0
azure-mgmt-advisor==1.0.1
azure-mgmt-applicationinsights==0.1.1
azure-mgmt-authorization==0.50.0
azure-mgmt-batch==5.0.1
azure-mgmt-batchai==2.0.0
azure-mgmt-billing==0.2.0
azure-mgmt-cdn==3.1.0
azure-mgmt-cognitiveservices==3.0.0
azure-mgmt-commerce==1.0.1
azure-mgmt-compute==4.6.2
azure-mgmt-consumption==2.0.0
azure-mgmt-containerinstance==1.5.0
azure-mgmt-containerregistry==2.8.0
azure-mgmt-containerservice==4.4.0
azure-mgmt-cosmosdb==0.4.1
azure-mgmt-datafactory==0.6.0
azure-mgmt-datalake-analytics==0.6.0
azure-mgmt-datalake-nspkg==3.0.1
azure-mgmt-datalake-store==0.5.0
azure-mgmt-datamigration==1.0.0
azure-mgmt-devspaces==0.1.0
azure-mgmt-devtestlabs==2.2.0
azure-mgmt-dns==2.1.0
azure-mgmt-eventgrid==1.0.0
azure-mgmt-eventhub==2.6.0
azure-mgmt-hanaonazure==0.1.1
azure-mgmt-iotcentral==0.1.0
azure-mgmt-iothub==0.5.0
azure-mgmt-iothubprovisioningservices==0.2.0
azure-mgmt-keyvault==1.1.0
azure-mgmt-loganalytics==0.2.0
azure-mgmt-logic==3.0.0
azure-mgmt-machinelearningcompute==0.4.1
azure-mgmt-managementgroups==0.1.0
azure-mgmt-managementpartner==0.1.1
azure-mgmt-maps==0.1.0
azure-mgmt-marketplaceordering==0.1.0
azure-mgmt-media==1.0.0
azure-mgmt-monitor==0.5.2
azure-mgmt-msi==0.2.0
azure-mgmt-network==2.7.0
azure-mgmt-notificationhubs==2.1.0
azure-mgmt-nspkg==3.0.2
azure-mgmt-policyinsights==0.1.0
azure-mgmt-powerbiembedded==2.0.0
azure-mgmt-rdbms==1.9.0
azure-mgmt-recoveryservices==0.3.0
azure-mgmt-recoveryservicesbackup==0.3.0
azure-mgmt-redis==5.0.0
azure-mgmt-relay==0.1.0
azure-mgmt-reservations==0.2.1
azure-mgmt-resource==2.2.0
azure-mgmt-scheduler==2.0.0
azure-mgmt-search==2.1.0
azure-mgmt-servicebus==0.5.3
azure-mgmt-servicefabric==0.2.0
azure-mgmt-signalr==0.1.1
azure-mgmt-sql==0.9.1
azure-mgmt-storage==2.0.0
azure-mgmt-subscription==0.2.0
azure-mgmt-trafficmanager==0.50.0
azure-mgmt-web==0.35.0
azure-nspkg==3.0.2
azure-servicebus==0.21.1
azure-servicefabric==6.3.0.0
azure-servicemanagement-legacy==0.20.7
azure-storage-blob==1.5.0
azure-storage-common==1.4.2
azure-storage-file==1.4.0
azure-storage-queue==1.4.0
certifi==2020.12.5
cffi==1.14.5
chardet==3.0.4
charset-normalizer==2.0.4
click==8.0.0
colorama==0.4.4
coverage==4.0.3
cryptography==3.4.7
cx-Freeze==6.6
cx-Logging==3.0
DateTime==4.3
distlib==0.3.1
Django==2.1.2
ecdsa==0.14.1
enum34==1.1.10
filelock==3.0.12
Flask==2.0.0
Flask-SQLAlchemy==2.5.1
future==0.18.2
greenlet==1.1.0
gunicorn==20.1.0
idna==2.7
importlib-metadata==4.4.0
iniconfig==1.1.1
intuit-oauth==1.1.1
isodate==0.6.0
itsdangerous==2.0.0
Jinja2==3.0.0
MarkupSafe==2.0.0
mock==4.0.3
msrest==0.6.21
msrestazure==0.6.4
numpy==1.20.1
oauthlib==3.1.0
packaging==20.9
pandas==1.2.2
pefile==2019.4.18
pluggy==0.13.1
py==1.10.0
pyasn1==0.4.8
pycparser==2.20
pyinstaller==4.2
pyinstaller-hooks-contrib==2020.11
PyJWT==2.1.0
pyodbc==4.0.30
pyparsing==2.4.7
pytest==6.2.4
pytest-cov==2.5.0
python-coveralls==2.9.3
python-dateutil==2.8.1
python-jose==3.2.0
pytz==2021.1
pywin32==300
pywin32-ctypes==0.2.0
PyYAML==5.4.1
requests==2.26.0
requests-oauthlib==1.3.0
rsa==4.7.2
six==1.15.0
SQLAlchemy==1.4.15
stripe==2.60.0
toml==0.10.2
urllib3==1.23
virtualenv==20.4.6
Werkzeug==2.0.0
zipp==3.4.1
zope.interface==5.4.0
